requirements.txt
mock == 4.0.3
pytest == 6.2.5
faust == 1.9.0
kafka-python == 2.0.2
